+++ packages/contracts/test/core/permission/integration-tests/parameter-scoping-oracle.ts
@@ -4,7 +4,7 @@ import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 
 import {
   TestParameterScopingPermissionOracle,
-  TestComponent,
+  TestPlugin,
   DAO,
 } from '../../../../typechain';
 import {customError} from '../../../test-utils/custom-error-helper';
@@ -14,7 +14,7 @@ const DO_SOMETHING_PERMISSION_ID = ethers.utils.id('DO_SOMETHING_PERMISSION');
 describe('TestParameterScopingOracle', function () {
   let signers: SignerWithAddress[];
   let parameterOracle: TestParameterScopingPermissionOracle;
-  let testComponent: TestComponent;
+  let testPlugin: TestPlugin;
   let managingDao: DAO;
   let ownerAddress: string;
   let expectedUnauthorizedError: string;
@@ -33,9 +33,9 @@ describe('TestParameterScopingOracle', function () {
     );
 
     // Deploy the component
-    const TestComponent = await ethers.getContractFactory('TestComponent');
-    testComponent = await TestComponent.deploy();
-    await testComponent.initialize(managingDao.address);
+    const TestPlugin = await ethers.getContractFactory('TestPlugin');
+    testPlugin = await TestPlugin.deploy();
+    await testPlugin.initialize(managingDao.address);
 
     // Deploy the oracle
     const ParameterOracle = await ethers.getContractFactory(
@@ -43,9 +43,9 @@ describe('TestParameterScopingOracle', function () {
     );
     parameterOracle = await ParameterOracle.deploy();
 
-    // Give signers[0] the `DO_SOMETHING_PERMISSION_ID` on the TestComponent
+    // Give signers[0] the `DO_SOMETHING_PERMISSION_ID` on the TestPlugin
     managingDao.grantWithOracle(
-      testComponent.address,
+      testPlugin.address,
       ownerAddress,
       DO_SOMETHING_PERMISSION_ID,
       parameterOracle.address
@@ -54,8 +54,8 @@ describe('TestParameterScopingOracle', function () {
     expectedUnauthorizedError = customError(
       'DaoUnauthorized',
       managingDao.address,
-      testComponent.address,
-      testComponent.address,
+      testPlugin.address,
+      testPlugin.address,
       ownerAddress,
       DO_SOMETHING_PERMISSION_ID
     );
@@ -67,7 +67,7 @@ describe('TestParameterScopingOracle', function () {
       let param2 = 1;
 
       expect(
-        await testComponent.callStatic.addPermissioned(param1, param2)
+        await testPlugin.callStatic.addPermissioned(param1, param2)
       ).to.be.equal(11);
     });
 
@@ -76,7 +76,7 @@ describe('TestParameterScopingOracle', function () {
       let param2 = 1;
 
       await expect(
-        testComponent.callStatic.subPermissioned(param1, param2)
+        testPlugin.callStatic.subPermissioned(param1, param2)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
 
@@ -85,7 +85,7 @@ describe('TestParameterScopingOracle', function () {
       let param2 = 1;
 
       await expect(
-        testComponent.callStatic.addPermissioned(param1, param2)
+        testPlugin.callStatic.addPermissioned(param1, param2)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
 
@@ -94,7 +94,7 @@ describe('TestParameterScopingOracle', function () {
       let param2 = 10;
 
       await expect(
-        testComponent.callStatic.addPermissioned(param1, param2)
+        testPlugin.callStatic.addPermissioned(param1, param2)
       ).to.be.revertedWith(expectedUnauthorizedError);
     });
   });